Standardizes search parameter fields and configures JSON storage.

+++ Backend/src/main/java/com/august/ScraperCar/dto/scraper/ScraperRequest.java
@@ -1,12 +1,14 @@
 package com.august.ScraperCar.dto.scraper;
 
+import java.math.BigDecimal;
+
 public record ScraperRequest(String nome,
                              int tipo,
                              int marca,
                              String string,
                              String versao,
-                             int valorinicio,
-                             int valorfim,
+                             BigDecimal valorinicio,
+                             BigDecimal valorfim,
                              int faixaano1,
                              int faixaano2,
                              String kminicio,
